requirements.txt
streamlit==1.38.0
requests==2.32.3
python-dotenv==1.0.1
pynini==2.1.5